Refactor files to apply norm in new expansion, creating new files and functions
