+++ Makefile
@@ -26,6 +26,7 @@ SOURCES = $(SRC_DIR)/main.c \
 					$(SRC_DIR)/parser/parser.c \
 					$(SRC_DIR)/parser/redir.c \
 					$(SRC_DIR)/expansion/alternatives.c \
+					$(SRC_DIR)/expansion/atoms.c \
 					$(SRC_DIR)/expansion/command.c \
 					$(SRC_DIR)/expansion/expansion.c \
 					$(SRC_DIR)/expansion/free.c \
