+++ source/expansion/command.c
@@ -6,7 +6,7 @@
 /*   By: migugar2 <migugar2@student.42madrid.com    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/08/16 16:52:47 by migugar2          #+#    #+#             */
-/*   Updated: 2025/08/30 15:29:46 by migugar2         ###   ########.fr       */
+/*   Updated: 2025/09/03 19:06:00 by migugar2         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -44,7 +44,8 @@ int	expand_cmd(t_shell *shell, t_cmd *cmd)
 	cmd->u_data.argv = NULL;
 	cmd->u_data.argv = convert_argv_to_array(&argv);
 	if (cmd->u_data.argv == NULL)
-		return (free_argv(&argv), free_exp_redirs(&cmd->redir), 1);
+		return (perror_malloc(), free_argv(&argv),
+			free_exp_redirs(&cmd->redir), 1);
 	cmd->count = argv.argc;
 	return (0);
 }
