+++ source/expansion/param.c
@@ -6,53 +6,71 @@
 /*   By: migugar2 <migugar2@student.42madrid.com    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/08/25 20:28:03 by migugar2          #+#    #+#             */
-/*   Updated: 2025/08/30 14:57:04 by migugar2         ###   ########.fr       */
+/*   Updated: 2025/09/03 19:08:36 by migugar2         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "minishell.h"
 
-/*
-t_param	*expand_param(t_shell *shell, t_seg *seg)
-{
-	t_param	*res;
-	char	*key;
 
-	res = new_param();
-	if (res == NULL)
-		return (NULL);
-	if (seg->slice.len == 1 && seg->slice.begin[0] == '?')
+// TODO: this is a really complex code for norm, but must be moved to a function in a verbose way
+int	multi_param(t_shell *shell, t_expand *build, char *val)
+{
+	size_t	i;
+	size_t	strt;
+
+	i = 0;
+	strt = 0;
+	while (val[i])
 	{
-		res->value = ft_itoa(shell->last_status);
-		res->len = ft_strlen(res->value);
-		return (res);
+		if (ft_isspace(val[i]) || val[i] == '*')
+		{
+			if (i > strt
+				&& append_atom(build, &val[strt], i - strt, ATOM_LIT) == 1)
+				return (1);
+			if (build->tail->head != NULL && ft_isspace(val[i]))
+				build->tail->flags |= BUILDF_FINISH;
+			while (ft_isspace(val[i]))
+				i++;
+			strt = i;
+			while (val[i] == '*')
+				i++;
+			if (i > strt && append_atom(build, NULL, i - strt, ATOM_WILD) == 1)
+				return (1);
+			strt = i;
+		}
+		else
+			i++;
 	}
-	key = malloc(sizeof(char) * (seg->slice.len + 1));
-	if (key == NULL)
-		return (free(res), NULL);
-	ft_strlcpy(key, seg->slice.begin, seg->slice.len + 1);
-	res->value = get_env_value(&shell->env_list, key);
-	free(key);
-	if (res->value == NULL)
-		res->len = 0;
-	else
-		res->len = ft_strlen(res->value);
-	return (res);
+	if (i > strt && val[strt] && !ft_isspace(val[strt]))
+		return (append_atom(build, &val[strt], i - strt, ATOM_LIT));
+	return (0);
+	(void)shell;
 }
 
-size_t	consume_param(t_exp *exp, t_seg *seg, char *cur)
+int	solve_param(t_shell *shell, t_seg *param, t_expand *build)
 {
-	t_param	*tmp;
-	size_t	len;
+	char	*key;
+	char	*value;
 
-	tmp = exp->head;
-	len = tmp->len;
-	if (tmp->value != NULL)
-		ft_strlcpy(cur, tmp->value, tmp->len + 1);
-	exp->head = exp->head->next;
-	free_param(&tmp, seg);
-	if (exp->head == NULL)
-		exp->tail = NULL;
-	return (len);
+	if (param->slice.len == 1 && param->slice.begin[0] == '?')
+	{
+		if (build->last_status == NULL)
+			build->last_status = ft_itoa(shell->last_status);
+		if (build->last_status == NULL)
+			return (1);
+		return (append_atom(build, build->last_status,
+				ft_strlen(build->last_status), ATOM_LIT));
+	}
+	key = malloc(sizeof(char) * (param->slice.len + 1));
+	if (key == NULL)
+		return (1);
+	ft_strlcpy(key, param->slice.begin, param->slice.len + 1);
+	value = get_env_value(&shell->env_list, key);
+	free(key);
+	if (value == NULL || value[0] == '\0')
+		return (0);
+	if (param->flags & SEGF_QUOTED || build->tail->flags & BUILDF_EQ)
+		return (append_atom(build, value, ft_strlen(value), ATOM_LIT));
+	return (multi_param(shell, build, value));
 }
-*/
