+++ source/expansion/wildcards.c
@@ -6,76 +6,24 @@
 /*   By: migugar2 <migugar2@student.42madrid.com    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/08/16 12:08:36 by migugar2          #+#    #+#             */
-/*   Updated: 2025/09/03 13:51:25 by migugar2         ###   ########.fr       */
+/*   Updated: 2025/09/03 18:44:03 by migugar2         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "minishell.h"
 
-t_atomtype	atom_peek(t_atom *atom, size_t offset, char *out)
-{
-	while (atom && offset >= atom->len)
-	{
-		offset -= atom->len;
-		atom = atom->next;
-	}
-	if (atom == NULL)
-		return (ATOM_END);
-	else if (atom->type == ATOM_WILD)
-	{
-		if (out != NULL)
-			*out = '*';
-		return (ATOM_WILD);
-	}
-	else if (atom->type == ATOM_LIT)
-	{
-		if (out != NULL)
-			*out = atom->value[offset];
-		return (ATOM_LIT);
-	}
-	return (ATOM_END);
-}
-
-void	atom_advance(t_atom **atom, size_t *offset)
-{
-	t_atom	*cur;
-	size_t	k;
-
-	cur = *atom;
-	k = *offset;
-	while (cur && k >= cur->len)
-	{
-		k -= cur->len;
-		cur = cur->next;
-	}
-	if (cur == NULL)
-	{
-		*atom = NULL;
-		*offset = k;
-		return ;
-	}
-	k++;
-	if (k >= cur->len)
-	{
-		cur = cur->next;
-		k = 0;
-	}
-	*offset = k;
-	*atom = cur;
-}
-
-int	wildcard_recursive(const char *str, t_atom *atom, size_t offset)
+int	wildcard_recursive(const char *str, t_atom *atom, size_t off)
 {
 	char	c;
 	t_atom	*newatom;
 	size_t	newoff;
 
-	if (atom_peek(atom, offset, &c) == ATOM_END || c == '\0')
+	if (atom_peek(atom, off, &c) == ATOM_END || c == '\0')
 		return (*str == '\0');
-	if (atom_peek(atom, offset, NULL) == ATOM_WILD)
+	if (atom_peek(atom, off, NULL) == ATOM_WILD)
 	{
 		newatom = atom;
-		newoff = offset;
+		newoff = off;
 		while (atom_peek(newatom, newoff, NULL) == ATOM_WILD)
 			atom_advance(&newatom, &newoff);
 		if (wildcard_recursive(str, newatom, newoff))
@@ -90,8 +38,7 @@ int	wildcard_recursive(const char *str, t_atom *atom, size_t offset)
 	}
 	if (*str == '\0' || *str != c)
 		return (0);
-	atom_advance(&atom, &offset);
-	return (wildcard_recursive(str + 1, atom, offset));
+	return (atom_advance(&atom, &off), wildcard_recursive(str + 1, atom, off));
 }
 
 int	wildcard_match(const char *str, t_atom **start, size_t *offset)
@@ -138,11 +85,10 @@ int	wildcard_solve(t_shell *shell, t_builder *builder, t_argv *argv)
 	t_atom			*start;
 	size_t			offset;
 
-	(void)shell;
 	if (start_wildcard(builder, &start, &offset) == 1)
 		return (0);
 	cur = opendir(".");
-	if (cur == NULL) // bash don't print error for permissions denied or something on wildcard match, only return and append the original input word
+	if (cur == NULL)
 	{
 		if (errno == ENOMEM)
 			return (perror_malloc(), 1);
@@ -159,6 +105,7 @@ int	wildcard_solve(t_shell *shell, t_builder *builder, t_argv *argv)
 		entry = readdir(cur);
 	}
 	return (closedir(cur), 0);
+	(void)shell;
 }
 
 int	expand_wildcards(t_shell *shell, t_builder *builder, t_argv *argv)
